backend/requirements.txt
Django>=3.1,<4.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-anymail[sendgrid]
python-caldav
django-redis
gunicorn
